pyproject.toml
[project]
name = "aquarela"
version = "0.1.0"
description = "Regatta instrument system for Nitro 80 Aquarela — Lake Lugano"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "python-can>=4.3",
    "pydantic>=2.0",
    "numpy>=1.26",
    "aiosqlite>=0.20",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
]

[tool.setuptools.packages.find]
include = ["aquarela*"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
